test(document): move solution-categories suites to a dedicated file (#2997)
